+++ go-app/app.go
@@ -927,6 +927,25 @@ func (a *App) OpenConfigDirectory() error {
 	return nil
 }
 
+func (a *App) SelectCommAPIKeyFile() (string, error) {
+	commapikey_path, err := runtime.OpenFileDialog(a.ctx, runtime.OpenDialogOptions{
+		Title: "Select the CommAPIKey.txt file",
+		Filters: []runtime.FileFilter{
+			{DisplayName: "CommAPIKey File", Pattern: "*.txt"},
+		},
+	})
+
+	if err != nil {
+		return "", fmt.Errorf("please select the CommAPIKey.txt file: %w", err)
+	}
+
+	if filepath.Base(commapikey_path) != "CommAPIKey.txt" {
+		return "", fmt.Errorf("please select the CommAPIKey.txt file")
+	}
+
+	return commapikey_path, nil
+}
+
 func (a *App) InstallTrainSimWorldMod() error {
 	tsw_exe_path, err := runtime.OpenFileDialog(a.ctx, runtime.OpenDialogOptions{
 		Title: "Select Train Sim World 5/6 executable within Binaries/Win64 (TrainSimWorld.exe)",
@@ -935,8 +954,8 @@ func (a *App) InstallTrainSimWorldMod() error {
 		return err
 	}
 
-	if filepath.Base(tsw_exe_path) != "TrainSimWorld.exe" {
-		return fmt.Errorf("you need to select the TrainSimWorld.exe file to install the mod")
+	if !strings.HasSuffix(filepath.ToSlash(tsw_exe_path), "Binaries/Win64/TrainSimWorld.exe") {
+		return fmt.Errorf("please select the TrainSimWorld.exe file in the game's Binaries/Win64 to install the mod")
 	}
 
 	var manifest ModAssets_Manifest